#1434 moved yes no prompt logic to new method added new UserAbortException moved getUserInput and yesNoPrompt to ValidationUtils changed error code from 0 to 255
