+++ cobigen-cli/cli/src/main/java/com/devonfw/cobigen/cli/exceptions/UserAbortException.java
@@ -1,0 +1,15 @@
+package com.devonfw.cobigen.cli.exceptions;
+
+import com.devonfw.cobigen.api.exception.CobiGenCancellationException;
+
+/**
+ * Exception is thrown if the generation process was cancelled by the user
+ */
+public class UserAbortException extends CobiGenCancellationException {
+
+  /**
+   * Default serial version UID.
+   */
+  private static final long serialVersionUID = 1L;
+
+}
